backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
litellm==1.65.6
python-dotenv==1.1.0
pydantic==2.11.3
httpx==0.28.1
